Refactor API into strings.ts and tests into strings_test.ts
